Moved the setDinoSprite method from the controller to the frameloader
