+++ phase-2/src/main/frameworkanddrivers/DogGameJPanel.java
@@ -1,4 +1,4 @@
-//package frameworkanddrivers;
+package frameworkanddrivers;
 
 import java.awt.*;
 import java.awt.event.KeyEvent;
